\name{alignment.summary}
\alias{alignment.summary}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ summary statistics of sequences }
\description{
  This function calculates the summary statistics of the alignments across genes.
}
\usage{
alignment.summary(seqfile,format="phylip")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{seqfile}{the input sequence files; one gene per file}
  \item{format}{format=phylip, fasta,nexus}
  
}

\author{ Liang Liu }

\keyword{ programming }
