\name{alignment.trim}
\alias{alignment.trim}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ alignment.trim sequences }
\description{
  This function trims gappy regions in the alignment.
}
\usage{
alignment.trim(path_trimal, inputfolder, outputfolder)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{path_trimal}{the full path of program Trimal}
  \item{inputfolder}{inputfolder contains the sequence files of original alignments in phylip format; one gene per file}
  \item{outputfolder}{outputfolder contains the sequence files of trimmed alignments}
}

\author{ Liang Liu }

\keyword{ programming }
