\name{file.beast2phylip}
\alias{file.beast2phylip}
\title{convert beast files to phylip files}
\description{
 This function converts beast xml files to phylip files 
}
\usage{
file.beast2phylip(beastfile)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{beastfile}{the beast xml file}
}
\author{ Liang Liu }

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ programming }
