\name{file.concatData}
\alias{file.concatData}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ concatenate sequences from multiple files }
\description{
  This function concatenates sequences from multiple files.
}
\usage{
file.concatData(inputfiles, confile)
}
\arguments{
  \item{inputfiles}{the sequence files to be concatenated}
  \item{confile}{the name of the output file to which the concatenated seuqences are saved}
}
\author{ Liang Liu }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ programming }
