\name{file.fasta2phylip}
\alias{file.fasta2phylip}
\title{convert fasta files to phylip files}
\description{
 This function converts fasta files to phylip files 
}
\usage{
file.fasta2phylip(inputfolder, outputfolder)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{inputfolder}{the folder of the fasta files}
  \item{outputfolder}{the folder of the phylip files }
}
\author{ Liang Liu }

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ programming }
