\name{file.phylip2nexus}
\alias{file.phylip2nexus}
\title{convert phylip files to nexus files}
\description{
 This function converts phylip files to nexus files 
}
\usage{
file.phylip2nexus(inputfolder, outputfolder)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{inputfolder}{the folder of the fasta files}
  \item{outputfolder}{the folder of the phylip files }
}
\author{ Liang Liu }

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ programming }
