\name{getncoal}
\alias{getncoal}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{internal function }
\description{
  This is an internal function for calculating the rannala and yang's formula
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ programming }
