\name{offspring.nodes.string}
\alias{offspring.nodes.string}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Find offspring nodes (internal use only) }
\description{
  The function returns a string of offspring nodes of \code{inode}. 
}

\usage{
offspring.nodes.string(inode, nodematrix, nspecies)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{inode}{ the node of which the the offspring nodes will be found by the function. }
  \item{nodematrix}{ the tree node matrix }
  \item{nspecies}{ the number of species }
}
\value{
  The function returns a string of offspring nodes of \code{inode}.
}
\author{ Liang Liu \email{lliu@uga.edu} }
\keyword{ programming }
