\name{parse.modeltest}
\alias{parse.modeltest}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{parse modeltest outputs}
\description{
  The function finds the selected models from the modeltest outputs 
}
\usage{
parse.modeltest(outputfile)
}
\arguments{
  \item{outputfile}{the modeltest output file}
}
\author{ Liang Liu \email{lliu@uga.edu} }

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ programming }
