\name{parse.phyml}
\alias{parse.phyml}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{parse phyml outputs}
\description{
  The function gets the parameter estimates from the phyml output files. 
}
\usage{
parse.phyml(phyml_stats_file="phyml_stats.txt")
}
\arguments{
  \item{phyml_stats_file}{phyml_stats_file}
}

\author{ Liang Liu \email{lliu@uga.edu} }

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ programming }
