\name{parse.raxml}
\alias{parse.raxml}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{parse raxml outputs}
\description{
  The function gets the parameter estimates from the raxml output files. 
}
\usage{
parse.raxml(raxml_info_file="RAxML_info.out")
}
\arguments{
  \item{raxml_info_file}{raxml_info_file}
}

\author{ Liang Liu \email{lliu@uga.edu} }

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ programming }
