\name{run.modeltest}
\alias{run.modeltest}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{run modeltest}
\description{
  The function runs jmodeltest2 
}
\usage{
run.modeltest(path_jmodeltest="./jmodeltest2/dist/jModelTest.jar", seqfile, nmodel=3, outputfile)
}
\arguments{
	\item{path_jmodeltest}{the path of jmodeltest2}
	\item{seqfile}{the input sequence file}
	\item{nmodel}{the number of models selected by jmodeltest2}
	\item{outputfile}{the result is saved to outputfile}
}
\author{ Liang Liu \email{lliu@uga.edu} }

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ programming }
