\name{sim.coal.mpest}
\alias{sim.coal.mpest}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Simulate gene trees from the mpest tree under the coalescent model}
\description{
This function can simulate gene trees from the mpest tree.
}
\usage{
sim.coal.mpest(mpest_tree,ngenetree)
}
\arguments{
  \item{mpest_tree}{the mpest tree}
 \item{ngenetree}{number of gene trees}

}
\value{
  The function returns the simulated gene trees.
}

\author{ Liang Liu \email{lliu@uga.edu} }

\keyword{ programming }
