\name{simnucleotide}
\alias{simnucleotide}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Intrinsic function used in sim.dna }
\description{
  The function simulates DNA sequences from a tree using the Jukes-Cantor model.
}
%- maybe also 'usage' for other objects documented here.
\author{ Liang Liu \email{lliu@uga.edu} }
\keyword{ programming }
