\name{site.summary}
\alias{site.summary}
\title{summarize alignments}
\description{
 This function summarizes alignments. 
}
\usage{
site.summary(sequence)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{sequence}{DNA alignment}
}
\author{ Liang Liu }

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ programming }
