\name{tree.probdist}
\alias{tree.probdist}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{tree.probdist}
\description{
  This function can be used to find the probability distribution of trees.
}
\usage{
tree.probdist(treefile, sumtreepath="sumtrees.py")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{treefile}{the tree file}
  \item{sumtreepath}{the full path of sumtrees.py}
}

\value{
  the function returns the probability distribution of trees.
}
\author{ Liang Liu }

\keyword{ programming }
