\name{triplenumber}
\alias{triplenumber}
\title{ Internal function }
\description{
This is an internal function used to calculate the loglikelihood of triples.
}
\usage{
triplenumber(dna)
}
\arguments{
  \item{dna}{ DNA sequences }
}
\details{This function is used to calculate triple likelihoods.
 }
\value{
  The function returns the number of triples.
}
\author{ Liang Liu \email{lliu@uga.edu} }
\keyword{IO}
\seealso{ \code{\link{write.subtree}}, \code{\link{read.tree.string}} }
