\name{tripleProb}
\alias{tripleProb}
\title{ Probability of a set of rooted triples }
\description{
  The function calculates the probability of a set of rooted triples.
}
\usage{
tripleProb(para)
}
\arguments{
  \item{para}{theta and gamma}
 }
\author{ Liang Liu \email{lliu@uga.edu} }
\keyword{programming}
