\name{unrooted.tree}
\alias{unrooted.tree}
\title{An example of unrooted trees}
\description{
  An example of unrooted trees
}
\usage{
data(unrooted.tree)
}
\author{ Liang Liu \email{lliu@uga.edu} }
\keyword{datasets}
\examples{
data(unrooted.tree)
read.tree.nodes(unrooted.tree[1])
}
