\name{site.summary}
\alias{site.summary}
\title{summarize alignments}
\description{
 This function summarizes alignments. 
}
\usage{
site.summary(sequence,is.cds)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{sequence}{DNA alignment}
  \item{is.cds}{If it is cds, GC content is separated by codon positions (12 vs 3)}
}
\author{ Liang Liu }

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ programming }
